chore: move FuseDriveStatus to virtual-drive module
