+++ src/backend/features/virtual-drive/FuseDriveStatus.ts
