+++ src/backend/features/virtual-drive/index.ts
@@ -1,2 +1,3 @@
 export { startFuseDaemonServer, stopFuseDaemonServer } from './services/server.service';
 export { startDaemon, stopDaemon } from './services/daemon.service';
+export { FuseDriveStatus } from './FuseDriveStatus';
